Revert "Revert "issue #2115 - Updated progress component to show progress bar instead of progress circle""
This reverts commit ce69cc956053d75b810f7f5382474c64991b3337.
