+++ src/helpers/projectHelper.js
@@ -1,7 +1,14 @@
 import _ from 'lodash'
-import { 
-  PROJECT_ROLE_CUSTOMER, PROJECT_ROLE_OWNER
+import moment from 'moment'
+
+import {
+  PROJECT_ROLE_CUSTOMER,
+  PROJECT_ROLE_OWNER,
+  PHASE_STATUS_ACTIVE,
+  PHASE_STATUS_COMPLETED,
 } from '../config/constants'
+
+import { formatNumberWithCommas } from './format'
 
 export const getProjectRoleForCurrentUser = ({currentUserId, project}) => {
   let role = null
@@ -60,4 +67,92 @@ export const setDuration = ({duration, status}) => {
     text = status === 'draft' ? 'Complete specification to get estimate' : 'Estimate not entered'
   }
   return { title, text, percent, type }
+}
+
+/**
+ * Format ProjectProgress props
+ *
+ * @param {Object} project project object
+ * @param {Array}  phases  project phases
+ *
+ * @return {Object} ProjectProgress props
+ */
+export function formatProjectProgressProps(project, phases) {
+  let actualDuration = 0
+  let now = new Date()
+  now = now && moment(now)
+
+  let totalProgress = 0
+
+  // phases where start date is set
+  const filteredPhases = _.filter(phases, (phase) => (phase.startDate))
+  filteredPhases.map((phase) => {
+    let progress = 0
+    // calculates days spent and day based progress for the phase
+    if (phase.startDate && phase.duration) {
+      const startDate = moment(phase.startDate)
+      const duration = now.diff(startDate, 'days')
+      if(duration >= 0) {
+        if(duration <= phase.duration) {
+          progress = (duration / phase.duration) * 100
+          actualDuration += duration
+        } else {
+          progress = 100
+          actualDuration += phase.duration
+        }
+      }
+    }
+    // override the progress use custom progress set by manager
+    if (phase.progress) {
+      progress = phase.progress
+    }
+    // override project progress if status is delivered
+    if (phase.status === PHASE_STATUS_COMPLETED) {
+      progress = 100
+    }
+    totalProgress += progress
+  })
+
+  const projectedDuration = _.sumBy(phases, 'duration') + phases.length
+
+  const labelDayStatus = `Day ${actualDuration} of ${projectedDuration}`
+
+  const activeOrCompletedPhases = _.filter(phases, (phase) => (
+    phase.status === PHASE_STATUS_ACTIVE || phase.status === PHASE_STATUS_COMPLETED)
+  )
+  const spentAmount = _.sumBy(activeOrCompletedPhases, 'spentBudget') || 0
+  const labelSpent = `Spent $${formatNumberWithCommas(spentAmount)}`
+  const progressPercent = phases.length > 0 ? Math.round(totalProgress/phases.length) : 0
+  const labelStatus = `${progressPercent}% done`
+
+  return {
+    labelDayStatus,
+    labelSpent,
+    labelStatus,
+    progressPercent,
+  }
+}
+
+/**
+ * Format ProjectProgress props for old projects
+ *
+ * @param {Object} project project object
+ *
+ * @return {Object} ProjectProgress props
+ */
+export function formatOldProjectProgressProps(project) {
+  const { duration: { actualDuration, projectedDuration } } = project
+  const { budget: { actualCost } } = project
+
+  const progressPercent = projectedDuration !== 0 ? Math.round(actualDuration/projectedDuration) : 0
+  const labelDayStatus = `Day ${actualDuration} of ${projectedDuration}`
+  const labelSpent = `Spent $${formatNumberWithCommas(actualCost)}`
+  const labelStatus = `${progressPercent}% done`
+
+  return {
+    labelDayStatus,
+    labelSpent,
+    labelStatus,
+    progressPercent,
+  }
 }
