+++ src/projects/detail/containers/ProjectInfoContainer.js
@@ -82,7 +82,7 @@ class ProjectInfoContainer extends React.Component {
 
   render() {
     const { duration } = this.state
-    const { project, currentMemberRole, isSuperUser } = this.props
+    const { project, currentMemberRole, isSuperUser, phases } = this.props
     let directLinks = null
     // check if direct links need to be added
     const isMemberOrCopilot = _.indexOf([PROJECT_ROLE_COPILOT, PROJECT_ROLE_MANAGER], currentMemberRole) > -1
@@ -118,6 +118,7 @@ class ProjectInfoContainer extends React.Component {
         <div className="sideAreaWrapper">
           <ProjectInfo
             project={project}
+            phases={phases}
             currentMemberRole={currentMemberRole}
             duration={duration}
             canDeleteProject={canDeleteProject}
@@ -148,7 +149,9 @@ class ProjectInfoContainer extends React.Component {
 
 ProjectInfoContainer.PropTypes = {
   currentMemberRole: PropTypes.string,
-  project: PropTypes.object.isRequired
+  phases: PropTypes.array,
+  project: PropTypes.object.isRequired,
+  isSuperUser: PropTypes.bool,
 }
 
 const mapDispatchToProps = { updateProject, deleteProject }
